NOJIRA Add test coverage to SIS import jobs

+++ nessie/jobs/import_degree_progress.py
@@ -37,6 +37,8 @@ from nessie.lib.util import get_s3_sis_api_daily_path, resolve_sql_template_stri
 
 class ImportDegreeProgress(BackgroundJob):
 
+    destination_schema = app.config['REDSHIFT_SCHEMA_STUDENT']
+
     def run(self, csids=None):
         if not csids:
             csids = [row['sid'] for row in get_all_student_ids()]
@@ -72,13 +74,14 @@ class ImportDegreeProgress(BackgroundJob):
             return False
 
         app.logger.info('Will copy S3 feeds into Redshift...')
-        query = resolve_sql_template_string(
+        if not redshift.execute(f'TRUNCATE {self.destination_schema}_staging.sis_api_degree_progress'):
+            app.logger.error('Error truncating old staging rows: aborting job.')
+            return False
+        if not redshift.copy_tsv_from_s3(f'{self.destination_schema}_staging.sis_api_degree_progress', s3_key):
+            app.logger.error('Error on Redshift copy: aborting job.')
+            return False
+        staging_to_destination_query = resolve_sql_template_string(
             """
-            TRUNCATE {redshift_schema_student}_staging.sis_api_degree_progress;
-            COPY {redshift_schema_student}_staging.sis_api_degree_progress
-                FROM '{loch_s3_sis_api_data_path}/degree_progress.tsv'
-                IAM_ROLE '{redshift_iam_role}'
-                DELIMITER '\\t';
             DELETE FROM {redshift_schema_student}.sis_api_degree_progress
                 WHERE sid IN (SELECT sid FROM {redshift_schema_student}_staging.sis_api_degree_progress);
             INSERT INTO {redshift_schema_student}.sis_api_degree_progress
@@ -86,7 +89,7 @@ class ImportDegreeProgress(BackgroundJob):
             TRUNCATE {redshift_schema_student}_staging.sis_api_profiles;
             """,
         )
-        if not redshift.execute(query):
+        if not redshift.execute(staging_to_destination_query):
             app.logger.error('Error on Redshift copy: aborting job.')
             return False
